+++ packages/terra-application-navigation/tests/jest/drawer-menu/DrawerMenuFooterButton.test.jsx
@@ -13,7 +13,6 @@ describe('DrawerMenuFooterButton', () => {
   it('should render with data props', () => {
     const shallowComponent = shallow(
       <DrawerMenuFooterButton
-        id="test-button-id"
         text="test-text"
       />,
     );