+++ packages/terra-application-navigation/tests/jest/extensions/Extension.test.jsx
@@ -5,7 +5,6 @@ describe('Extension', () => {
   it('should render with required prop data', () => {
     const shallowComponent = shallow(
       <Extension
-        id="test-extension-id"
         icon={<span>my icon</span>}
         text="my text"
       />,