+++ packages/terra-application-navigation/tests/jest/extensions/Extensions.test.jsx
@@ -17,9 +17,7 @@ describe('Extensions', () => {
       <Extensions.WrappedComponent
         intl={mockIntl}
         activeBreakpoint="small"
-        extensionItems={[{
-          id: 'test-extension-id', text: 'test-text', icon: <span>my icon</span>, key: 'my-test-key',
-        }]}
+        extensionItems={[{ text: 'test-text', icon: <span>my icon</span>, key: 'my-test-key' }]}
         notifications={{ myComponent: 3 }}
         onSelect={jest.fn()}
       />,