+++ packages/terra-application-navigation/tests/jest/tabs/Tab.test.jsx
@@ -15,7 +15,6 @@ describe('Tab', () => {
   it('should render prop data', () => {
     const shallowComponent = shallow(
       <Tab
-        id="test-tab-id"
         notificationCount={1}
         text="my text"
         isActive